+++ client/homebrew/pages/userPage/brewItem/brewItem.jsx
@@ -26,7 +26,7 @@ const BrewItem = createClass({
 			if(!confirm('Are you REALLY sure? You will lose editor access to this document.')) return;
 		}
 
-		request.get(`/api/remove/${this.props.brew.editId}`)
+		request.delete(`/api/${this.props.brew.editId}`)
 			.send()
 			.end(function(err, res){
 				location.reload();