+++ client/homebrew/pages/userPage/userPage.jsx
@@ -42,15 +42,6 @@ const UserPage = createClass({
 		});
 	},
 
-	renderPrivateBrews : function(privateBrews){
-		if(!privateBrews || !privateBrews.length) return;
-
-		return [
-			<h1>{this.props.username}'s unpublished brews</h1>,
-			this.renderBrews(privateBrews)
-		];
-	},
-
 	render : function(){
 		const brews = this.getSortedBrews();
 
@@ -64,9 +55,14 @@ const UserPage = createClass({
 
 			<div className='content'>
 				<div className='phb'>
-					<h1>{this.props.username}'s brews</h1>
-					{this.renderBrews(brews.published)}
-					{this.renderPrivateBrews(brews.private)}
+					<div>
+						<h1>{this.props.username}'s brews</h1>
+						{this.renderBrews(brews.published)}
+					</div>
+					<div>
+						<h1>{this.props.username}'s unpublished brews</h1>
+						{this.renderBrews(brews.private)}
+					</div>
 				</div>
 			</div>
 		</div>;